refac: renderer class
