+++ src/lib/constants/entity.ts
@@ -1,0 +1,21 @@
+export const ENTITY = {
+	POSITION: {
+		// 엔티티가 스폰되는 시작 위치
+		DEFAULT_START_X: 50,
+		DEFAULT_START_Y: 50
+	},
+	// 엔티티의 첫번째 타이틀 행 정보
+	TITLE_ROW: {
+		DEFAULT_HEIGHT: 30,
+		DEFAULT_FONT_SIZE: 20,
+		DEFAULT_COLOR: 'black',
+		DEFAULT_STROKE_COLOR: 'black',
+		DEFAULT_STROKE_WIDTH: 4
+	},
+	ROW: {
+		// 엔티티 나머지 행의 세로 크기
+		DEFAULT_HEIGHT: 25,
+		DEFAULT_WIDTH: 300,
+		DEFAULT_FONT_SIZE: 17
+	}
+};
