+++ src/lib/constants/note.ts
@@ -1,0 +1,14 @@
+export const NOTE = {
+	DEFAULT_WIDTH: 200,
+	DEFAULT_HEIGHT: 100,
+	DEFAULT_COLOR: 'grey',
+	DEFAULT_STROKE_COLOR: 'grey',
+	DEFAULT_STROKE_WIDTH: 4,
+	DEFAULT_START_X: 50,
+	DEFAULT_START_Y: 50,
+	INPUT: {
+		DEFAULT_WIDTH: 200,
+		DEFAULT_HEIGHT: 100,
+		DEFAULT_FONT_SIZE: 17
+	}
+};
